+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.0.2
+
+- InstantWipe no longer crashes the game on player death
+- Disabled Hardcore (for now) because it caused lots of bugs
+- Set MapSizeMultiplier default value to 1
+
 ## v1.0.1
 
 - Made the mod actually work properly